♻️ Refactor interface
